updating horizontal rule and moving code to molecu
