+++ components/atoms/HorizontalRule.js
@@ -1,0 +1,3 @@
+export function HorizontalRule() {
+  return <hr className="h-1 bg-hr-red-bar"></hr>;
+}
